+++ zigzag.legendweathers/CHANGELOG.md
@@ -5,6 +5,7 @@
     - Added `Terrain effects volume` config, this can be used to customize the volume of the weather's terrain effects
     - Added `Ambience music volume` config, this can be used to customize the volume of the weather's ambient music
     - Added `Ambience music type` config, you can use this to replace the weather's ambient music by the vanilla Eclipsed music if wanted
+    - Added `Blood Stone spawning blacklist` config, this can be filled by the host with enemy names that will never spawn Blood Stones when they are killed
 - **Majora Moon**
     - f